requirements.txt
Flask==2.0.3
gunicorn==19.9.0
line-bot-sdk==1.14.0
flask-sqlalchemy==2.5.1
apscheduler==3.8.1
datetime==4.3
beautifulsoup4==4.10.0
pymysql==0.9.3
Flask-Login==0.5.0
pandas-ta==0.3.14b0
pandas==1.3.5
pandas-datareader==0.10.0
plotly==5.6.0
numpy==1.21.5
finta==1.3
yfinance==0.1.70
jupyterlab==3
ipywidgets==7.6
matplotlib==3.5.1
flask-paginate==2022.1.8
Backtesting==0.3.3
